Move public-register CMS plugins and templates from danceschool.register to danceschool.core (Step 4)
Plugin templates moved via git mv (history preserved):
  register/templates/register/plugins/public_event_register.html
    → core/templates/core/plugins/public_event_register.html
  register/templates/register/plugins/public_register_nav.html
    → core/templates/core/plugins/public_register_nav.html

Plugin classes extracted from register/cms_plugins.py (history traceable via
git log --find-copies-harder on the original file) into core/cms_plugins.py:
  PublicRegisterNavPlugin, PublicRegisterEventChoiceInline,
  PublicRegisterEventPlugin

render_template paths updated from 'register/plugins/...' to 'core/plugins/...'.
SortableInlineAdminMixin and ensure_localtime imports added to core/cms_plugins.py.
register/cms_plugins.py stripped of all public-register code and imports.
