+++ danceschool/core/templates/core/plugins/public_event_register.html
