+++ danceschool/core/templates/core/plugins/public_register_nav.html
